Wait a few seconds after arduino connection
Some arduinos are resetting after being connected to. This result in the
'V' command to timeout (ie sent too early while arduino is not ready to
handle it). Wait 2 seconds before sending the command so that the
arduino is ready to answer. This commit fixes issues #47 which I also
encountered.

Signed-off-by: Clément Léger <clement@clement-leger.fr>